+++ app/gui/src/project-view/components/GraphEditor/widgets/WidgetTableEditor/TableHeader.vue
@@ -1,23 +1,35 @@
 <script lang="ts">
+import SvgButton from '@/components/SvgButton.vue'
+import { provideTooltipRegistry, TooltipRegistry } from '@/providers/tooltipState'
 import type { IHeaderParams } from 'ag-grid-community'
-import { ref, watch } from 'vue'
+import { computed, ref, watch } from 'vue'
+
+export interface HeaderEditHandlers {
+  /** Setter called when column name is changed by the user. */
+  nameSetter: (newName: string) => void
+  onHeaderEditingStarted?: (stop: (cancel: boolean) => void) => void
+  onHeaderEditingStopped?: () => void
+}
 
 /**
  * Parameters recognized by this header component.
  *
  * They are set through `headerComponentParams` option in AGGrid column definition.
  */
-export interface HeaderParams {
-  /** Setter called when column name is changed by the user. */
-  nameSetter?: (newName: string) => void
+export type HeaderParams = {
   /**
-   * Column is virtual if it is not represented in the AST. Such column might be used
-   * to create new one.
+   * AgGrid mounts header components as separate "App", so we don't have access to any context.
+   * Threfore the tooltip registry must be provided by props.
    */
-  virtualColumn?: boolean
-  onHeaderEditingStarted?: (stop: (cancel: boolean) => void) => void
-  onHeaderEditingStopped?: () => void
-}
+  tooltipRegistry: TooltipRegistry
+} & (
+  | {
+      type: 'astColumn'
+      editHandlers: HeaderEditHandlers
+    }
+  | { type: 'newColumn'; newColumnRequested: () => void }
+  | { type: 'rowIndexColumn ' }
+)
 </script>
 
 <script setup lang="ts">
@@ -25,17 +37,23 @@ const props = defineProps<{
   params: IHeaderParams & HeaderParams
 }>()
 
+/** Re-provide tooltipRegistry. See `tooltipRegistry` docs in {@link HeaderParams} */
+provideTooltipRegistry.provideConstructed(props.params.tooltipRegistry)
+
 const editing = ref(false)
 const inputElement = ref<HTMLInputElement>()
+const editHandlers = computed(() =>
+  props.params.type === 'astColumn' ? props.params.editHandlers : undefined,
+)
 
 watch(editing, (newVal) => {
   if (newVal) {
-    props.params.onHeaderEditingStarted?.((cancel: boolean) => {
+    editHandlers.value?.onHeaderEditingStarted?.((cancel: boolean) => {
       if (cancel) editing.value = false
       else acceptNewName()
     })
   } else {
-    props.params.onHeaderEditingStopped?.()
+    editHandlers.value?.onHeaderEditingStopped?.()
   }
 })
 
@@ -48,33 +66,47 @@ watch(inputElement, (newVal, oldVal) => {
 })
 
 function acceptNewName() {
+  if (editHandlers.value == null) {
+    console.error("Tried to accept header new name where it's not editable!")
+    return
+  }
   if (inputElement.value == null) {
     console.error('Tried to accept header new name without input element!')
     return
   }
-  props.params.nameSetter?.(inputElement.value.value)
+  editHandlers.value.nameSetter(inputElement.value.value)
   editing.value = false
 }
 
-function onMouseClick() {
-  if (!editing.value && props.params.nameSetter != null) {
+function onMouseClick(event: MouseEvent) {
+  if (!editing.value && props.params.type === 'astColumn') {
     editing.value = true
+    event.stopPropagation()
   }
 }
 
 function onMouseRightClick(event: MouseEvent) {
   if (!editing.value) {
     props.params.showColumnMenuAfterMouseClick(event)
+    event.preventDefault()
+    event.stopPropagation()
   }
 }
 </script>
 
 <template>
+  <SvgButton
+    v-if="params.type === 'newColumn'"
+    class="addColumnButton"
+    name="add"
+    title="Add new column"
+    @click.stop="params.newColumnRequested()"
+  />
   <div
+    v-else
     class="ag-cell-label-container"
     role="presentation"
     @pointerdown.stop
-    @click.stop
     @click="onMouseClick"
     @click.right="onMouseRightClick"
   >
@@ -93,14 +125,18 @@ function onMouseRightClick(event: MouseEvent) {
       <span
         v-else
         class="ag-header-cell-text"
-        :class="{ virtualColumn: params.virtualColumn === true }"
+        :class="{ virtualColumn: params.type !== 'astColumn' }"
         >{{ params.displayName }}</span
       >
     </div>
   </div>
 </template>
 
-<style>
+<style scoped>
+.addColumnButton {
+  margin-left: 10px;
+}
+
 .virtualColumn {
   color: rgba(0, 0, 0, 0.5);
 }