+++ app/gui/src/project-view/providers/tooltipState.ts
@@ -31,6 +31,7 @@ const { provideFn, injectFn } = createContextStore('tooltip registry', () => {
       return set ? last(set) : undefined
     },
     registerTooltip(slot: Ref<Slot | undefined>) {
+      console.log('REGISTER TOOLTIP')
       const entry: TooltipEntry = {
         contents: slot,
         key: Symbol(),